Move Candidate + retrieval functions to separate module. Fix other, small issues.
