+++ spacy/kb/__init__.py
@@ -1,2 +1,3 @@
-from .kb import KnowledgeBase, Candidate, get_candidates, get_candidates_batch
+from .kb import KnowledgeBase
 from .kb_in_memory import InMemoryLookupKB
+from .candidate import Candidate, get_candidates, get_candidates_batch
